Fixed bugs in Signup Use Case and Login View
