+++ src/main/java/app/AppBuilder.java
@@ -9,9 +9,9 @@ import interface_adaptor.Menu.MenuViewModel;
 import interface_adaptor.Menu.StarRateController;
 import interface_adaptor.Menu.StarRatePresenter;
 import interface_adaptor.ViewManagerModel;
-import interface_adaptor.signup.SignupController;
-import interface_adaptor.signup.SignupPresenter;
-import interface_adaptor.signup.SignupViewModel;
+import interface_adaptor.Signup.SignupController;
+import interface_adaptor.Signup.SignupPresenter;
+import interface_adaptor.Signup.SignupViewModel;
 import log_in.LoginInputBoundary;
 import log_in.LoginInteractor;
 import log_in.LoginOutputBoundary;
