+++ src/main/java/interface_adaptor/Signup/SignupController.java
@@ -1,16 +1,34 @@
 package interface_adaptor.Signup;
 
-import sign_up.SignupInputBoundary;
-import sign_up.SignupInputData;
+import signup.SignupInputBoundary;
+import signup.SignupInputData;
 
 public class SignupController {
-    private final SignupInputBoundary interactor;
 
-    public SignupController(SignupInputBoundary interactor) {
-        this.interactor = interactor;
+    private final SignupInputBoundary userSignupUseCaseInteractor;
+
+    public SignupController(SignupInputBoundary userSignupUseCaseInteractor) {
+        this.userSignupUseCaseInteractor = userSignupUseCaseInteractor;
     }
 
-    public void execute(String username, String password) {
-        interactor.execute(new SignupInputData(username, password));
+    /**
+     * Executes the Signup Use Case.
+     *
+     * @param username  the username to sign up
+     * @param password1 the password
+     * @param password2 the password repeated
+     */
+    public void execute(String username, String password1, String password2) {
+        final SignupInputData signupInputData = new SignupInputData(
+                username, password1, password2);
+
+        userSignupUseCaseInteractor.execute(signupInputData);
+    }
+
+    /**
+     * Executes the "switch to LoginView" Use Case.
+     */
+    public void switchToLoginView() {
+        userSignupUseCaseInteractor.switchToLoginView();
     }
 }
