+++ src/main/java/interface_adaptor/Signup/SignupState.java
@@ -1,8 +1,67 @@
 package interface_adaptor.Signup;
 
 public class SignupState {
-    private String message = "";
+    private String username = "";
+    private String usernameError;
+    private String password = "";
+    private String passwordError;
+    private String repeatPassword = "";
+    private String repeatPasswordError;
 
-    public String getMessage() { return message; }
-    public void setMessage(String msg) { this.message = msg; }
+    public String getUsername() {
+        return username;
+    }
+
+    public String getUsernameError() {
+        return usernameError;
+    }
+
+    public String getPassword() {
+        return password;
+    }
+
+    public String getPasswordError() {
+        return passwordError;
+    }
+
+    public String getRepeatPassword() {
+        return repeatPassword;
+    }
+
+    public String getRepeatPasswordError() {
+        return repeatPasswordError;
+    }
+
+    public void setUsername(String username) {
+        this.username = username;
+    }
+
+    public void setUsernameError(String usernameError) {
+        this.usernameError = usernameError;
+    }
+
+    public void setPassword(String password) {
+        this.password = password;
+    }
+
+    public void setPasswordError(String passwordError) {
+        this.passwordError = passwordError;
+    }
+
+    public void setRepeatPassword(String repeatPassword) {
+        this.repeatPassword = repeatPassword;
+    }
+
+    public void setRepeatPasswordError(String repeatPasswordError) {
+        this.repeatPasswordError = repeatPasswordError;
+    }
+
+    @Override
+    public String toString() {
+        return "SignupState{"
+                + "username='" + username + '\''
+                + ", password='" + password + '\''
+                + ", repeatPassword='" + repeatPassword + '\''
+                + '}';
+    }
 }
