+++ src/main/java/interface_adaptor/Signup/SignupViewModel.java
@@ -1,22 +1,22 @@
 package interface_adaptor.Signup;
 
-import interface_adaptor.ViewManagerModel;
 import interface_adaptor.ViewModel;
 
 public class SignupViewModel extends ViewModel<SignupState> {
 
-    private ViewManagerModel viewManagerModel;
+    public static final String TITLE_LABEL = "Sign Up";
+    public static final String USERNAME_LABEL = "Choose username";
+    public static final String PASSWORD_LABEL = "Choose password";
+    public static final String REPEAT_PASSWORD_LABEL = "Enter password again";
+
+    public static final String SIGNUP_BUTTON_LABEL = "Sign up";
+    public static final String CANCEL_BUTTON_LABEL = "Cancel";
+
+    public static final String TO_LOGIN_BUTTON_LABEL = "Go to Login";
 
     public SignupViewModel() {
-        super("signup");
+        super("sign up");
         setState(new SignupState());
     }
 
-    public void setViewManagerModel(ViewManagerModel vmm) {
-        this.viewManagerModel = vmm;
-    }
-
-    public ViewManagerModel getViewManagerModel() {
-        return viewManagerModel;
-    }
 }
