--- src/main/java/interface_adaptor/signup/SignupController.java
@@ -1,34 +1,0 @@
-package interface_adaptor.signup;
-
-import signup.SignupInputBoundary;
-import signup.SignupInputData;
-
-public class SignupController {
-
-    private final SignupInputBoundary userSignupUseCaseInteractor;
-
-    public SignupController(SignupInputBoundary userSignupUseCaseInteractor) {
-        this.userSignupUseCaseInteractor = userSignupUseCaseInteractor;
-    }
-
-    /**
-     * Executes the Signup Use Case.
-     *
-     * @param username  the username to sign up
-     * @param password1 the password
-     * @param password2 the password repeated
-     */
-    public void execute(String username, String password1, String password2) {
-        final SignupInputData signupInputData = new SignupInputData(
-                username, password1, password2);
-
-        userSignupUseCaseInteractor.execute(signupInputData);
-    }
-
-    /**
-     * Executes the "switch to LoginView" Use Case.
-     */
-    public void switchToLoginView() {
-        userSignupUseCaseInteractor.switchToLoginView();
-    }
-}
