--- src/main/java/interface_adaptor/signup/SignupViewModel.java
@@ -1,23 +1,0 @@
-package interface_adaptor.signup;
-
-import interface_adaptor.ViewManagerModel;
-import interface_adaptor.ViewModel;
-
-public class SignupViewModel extends ViewModel<SignupState> {
-
-    public static final String TITLE_LABEL = "Sign Up";
-    public static final String USERNAME_LABEL = "Choose username";
-    public static final String PASSWORD_LABEL = "Choose password";
-    public static final String REPEAT_PASSWORD_LABEL = "Enter password again";
-
-    public static final String SIGNUP_BUTTON_LABEL = "Sign up";
-    public static final String CANCEL_BUTTON_LABEL = "Cancel";
-
-    public static final String TO_LOGIN_BUTTON_LABEL = "Go to Login";
-
-    public SignupViewModel() {
-        super("sign up");
-        setState(new SignupState());
-    }
-
-}
